Moved analyze to report.py
