+++ python/refactored/ITC.py
@@ -46,144 +46,10 @@ import Constants
 import numpy
 import scipy.stats
 import pymc
-from report import Report
+from report import Report, analyze
 from experiments import Injection, Experiment
 from instruments import VPITC
 from models import RescalingStep, TwoComponentBindingModel
-
-
-#=============================================================================================
-# Report class
-#=============================================================================================
-
-  
-def analyze(name, experiment):
-    
-    #=============================================================================================
-    # Make plots
-    #=============================================================================================
-    
-    # Create a LaTeX report file.
-    report = Report([experiment])
-
-    # Plot the raw measurements of differential power versus time and the enthalpogram.
-    import pylab
-    pylab.figure()
-
-    fontsize = 8
-
-    # Plot differential power versus time.
-    pylab.subplot(411)
-
-    # plot baseline fit
-    pylab.hold(True)
-    pylab.plot(experiment.filter_period_end_time / Units.s, experiment.baseline_power / (Units.ucal/Units.s), 'g-') # plot baseline fit
-
-    # differential power
-    pylab.plot(experiment.filter_period_end_time / Units.s, experiment.differential_power / (Units.ucal/Units.s), 'k.', markersize=1)
-    # plot red vertical line to mark injection times
-    pylab.hold(True)    
-    [xmin, xmax, ymin, ymax] = pylab.axis()
-
-    for injection in experiment.injections:
-        last_index = injection['first_index'] # timepoint at start of syringe injection
-        t = experiment.filter_period_end_time[last_index] / Units.s
-        pylab.plot([t, t], [ymin, ymax], 'r-')    
-    pylab.hold(False)
-    xlabel = pylab.xlabel('time / s')
-    xlabel.set_fontsize(fontsize)
-    ylabel = pylab.ylabel('differential power / ucal/s')
-    ylabel.set_fontsize(fontsize)
-
-    ax = pylab.gca()
-    for tick in ax.xaxis.get_major_ticks():
-        tick.label1.set_fontsize(fontsize)
-    for tick in ax.yaxis.get_major_ticks():
-        tick.label1.set_fontsize(fontsize)
-
-    # title plot
-    title = pylab.title(name)
-    title.set_fontsize(fontsize)
-    
-    # Plot enthalpogram.
-    pylab.subplot(412)
-    pylab.hold(True)
-    for injection in experiment.injections:
-        # determine initial and final samples for injection i
-        first_index = injection['first_index'] # index of timepoint for first filtered differential power measurement
-        last_index  = injection['last_index']  # index of timepoint for last filtered differential power measurement
-        # determine time at end of injection period
-        t = experiment.filter_period_end_time[last_index] / Units.s
-        # plot a point there to represent total heat evolved in injection period
-        y = injection['evolved_heat'] / Units.ucal
-        pylab.plot([t, t], [0, y], 'k-')
-        # label injection
-        pylab.text(t, y, '%d' % injection['number'], fontsize=6)        
-    # adjust axes to match first plot
-    [xmin_new, xmax_new, ymin, ymax] = pylab.axis()
-    pylab.axis([xmin, xmax, ymin, ymax])    
-    pylab.hold(False)
-    #pylab.title('evolved heat per injection')
-    xlabel = pylab.xlabel('time / s')
-    xlabel.set_fontsize(fontsize)
-    ylabel = pylab.ylabel('evolved heat / ucal')
-    ylabel.set_fontsize(fontsize)
-    # plot zero
-    pylab.hold(True)
-    pylab.plot(experiment.filter_period_end_time / Units.s, 0.0*experiment.filter_period_end_time / (Units.ucal/Units.s), 'g-') # plot zero line
-
-
-    ax = pylab.gca()
-    for tick in ax.xaxis.get_major_ticks():
-        tick.label1.set_fontsize(fontsize)
-    for tick in ax.yaxis.get_major_ticks():
-        tick.label1.set_fontsize(fontsize)
-                    
-    # Plot cell temperature.
-    pylab.subplot(413)
-    pylab.hold(True)
-    pylab.plot(experiment.filter_period_end_time / Units.s, experiment.cell_temperature/Units.K - Constants.absolute_zero, 'r.', markersize=1)
-    # adjust axes to match first plot
-    [xmin_new, xmax_new, ymin, ymax] = pylab.axis()
-    pylab.axis([xmin, xmax, ymin, ymax])    
-    pylab.hold(False)
-    #pylab.title('evolved heat per injection')
-    xlabel = pylab.xlabel('time / s')
-    xlabel.set_fontsize(fontsize)
-    ylabel = pylab.ylabel('cell temperature / C')
-    ylabel.set_fontsize(fontsize)
-
-    ax = pylab.gca()
-    for tick in ax.xaxis.get_major_ticks():
-        tick.label1.set_fontsize(fontsize)
-    for tick in ax.yaxis.get_major_ticks():
-        tick.label1.set_fontsize(fontsize)
-
-    # Plot adiabatic jacket temperature.
-    pylab.subplot(414)
-    pylab.hold(True)
-    pylab.plot(experiment.filter_period_end_time / Units.s, experiment.jacket_temperature/Units.K - Constants.absolute_zero, 'b.', markersize=1)    
-    # adjust axes to match first plot
-    [xmin_new, xmax_new, ymin, ymax] = pylab.axis()
-    pylab.axis([xmin, xmax, ymin, ymax])    
-    pylab.hold(False)
-    #pylab.title('evolved heat per injection')
-    xlabel = pylab.xlabel('time / s')
-    xlabel.set_fontsize(fontsize)
-    ylabel = pylab.ylabel('jacket temperature / C')
-    ylabel.set_fontsize(fontsize)
-    
-    ax = pylab.gca()
-    for tick in ax.xaxis.get_major_ticks():
-        tick.label1.set_fontsize(fontsize)
-    for tick in ax.yaxis.get_major_ticks():
-        tick.label1.set_fontsize(fontsize)
-
-    # show plot
-    #pylab.show()
-    pylab.savefig('%s.pdf' % name, orientation='landscape', papertype='letter', format='pdf')
-
-    return
 
 
 def compute_statistics(x_t):
